removed back buttons from detail screens and reorganized folder structure for screens
